requirements.txt
# OncoScope Requirements
# Core backend dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.8.2
pydantic-settings==2.10.1
python-multipart==0.0.6

# AI/ML dependencies
requests==2.31.0
aiohttp==3.9.1
numpy==1.24.3
torch==2.1.0
transformers==4.35.0
datasets==2.14.0
trl==0.7.7

# Database
sqlalchemy==2.0.23
aiosqlite==0.19.0

# Utilities
python-dotenv==1.0.0
rich==13.7.0
pytest==7.4.3
pytest-asyncio==0.21.1

# UI (optional)
gradio==4.44.1

# Data processing
pandas==2.1.3
biopython==1.81  # For genetic sequence handling

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.14